Explicitly define "nacl_defines" when OS is "ios"
Plan for building for arm64 on iOS is to set "target_arch" to correct
value (currently default to "ia32"). As "nacl_defines" needs to be
defined even if "disable_nacl" is set to "1", define it to an empty list
when generating project for iOS so that changing the "target_arch"
won't cause build failures.

BUG=314720
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/158873003