Need to also scale scroll offset, not just the scroll rect.
When not using the two-viewport model, the page scale layer ends up
coinciding with the viewport clip layer. Prior to this CL, in this
case the page scale was being applied to the scroll layer rect, but
not to the current scroll offset, when calculating the scrollbar
size/positioning. This CL apples the scale to the offset as well.

BUG=340631

Review URL: https://codereview.chromium.org/159123002