sync: Stop integration tests on timeout
The integration tests' Await functions tend to return false on timeout
and hope that their callers will ASSERT_ on their return value.  That
doesn't always happen, though.  There are some old tests that expect the
Await functions to return false (for non-timeout related reasons) or
don't bother to check.  One such set of tests led to crbug.com/341215.

This change will force the tests to halt when the timeout occurs.  We
hope this will expose pre-existing issues that are hidden by the current
soft-timeout behavior.

Killing tests with a CHECK() is normally a bad idea.  It can prevent
other tests in the suite from running.  In this case we can get away
with it, since the integration test runner won't halt if one of its
tests is killed.

BUG=341301

Review URL: https://codereview.chromium.org/156213006