cc: Prevent usage of rasterize on-demand with delegating renderer.
PictureDrawQuads are not yet supported by the delegating renderer.
Generating this type of quads for on-demand raster of tiles cause
invalid IPC to be sent to the parent compositor and the renderer
to be killed.

This makes us activate with low-res or checkerboard tiles until
PictureDrawQuads are properly supported by the delegating renderer.

BUG=340695
TEST=cc_unittests --gtest_filter=*RasterizeOnDemand*

Review URL: https://codereview.chromium.org/157743007