Cast receiver hits a DCHECK if sending too fast
UdpTransport used by cast_receiver_app disregard the send callback.
If receiver is sending too fast it will send even if there is a pending
write. This will hit a DCHECK in UDPSocketLibevent.

The fix is to drop packets if there is a pending write.

BUG=b/12611188

Review URL: https://codereview.chromium.org/159053003