Change the mock web theme to not depend on the order of the states of the control.
We were figuring out which color to draw a control based on the state, but
doing so by using the state as an indexed lookup into an array (making the
ordering of the enums in the state significant).

This patch changes the lookup to a function call w/ a switch-based dispatch;
slightly slower, but should still be fast, and this makes the values of
the enum not matter.

R=jamesr@chromium.org
BUG=327471

Review URL: https://codereview.chromium.org/157983007