Move silk and fast_path benchmarks into their proper places
We are try name benchmarks be named by the measurement name. This changes silk
and fast path to keep that convention. But, in deference to folks knowing the
old names and buildbot pointing at the old names, this patch includes the old
names for backcompat.

R=ernstm
NOTRY=True

Review URL: https://codereview.chromium.org/148663006