Create tool that will asynchronously poll file content.
When run with 2 or more arguments the file_poller tool will open a port on
the device, print it on its standard output and then start collect files content. The first argument is the polling rate in Hz, and the following
arguments are file to poll.
When run with the port of an already running file_poller, the tool will
contact the first instance, retrieve the sample and print those on its
standard output. This will also terminate the first instance.

NOTRY=true
R=tonyg@chromium.org,pliard@chromium.org

Review URL: https://codereview.chromium.org/152093004