gesture nav: Fix the UX for some navigation edge cases.
The renderer does not always send the 'first visual non-empty paint' message
to the browser for navigations (e.g. when doing in-page navigation). As a
result, the screenshot overlay never hides in such cases. So force a repaint of
the page right after the page has completely loaded if the 'first paint' message
hasn't been received by then, and hide the overlay when the paint happens.

This also adds some test coverage for the overlay (including for this change).

BUG=342311
R=sky@chromium.org
TBR=piman@ for content_tets.gypi

Review URL: https://codereview.chromium.org/138253008