[invalidations] Added logging to about:invalidations
This patch is a follow-up to issue 144093005.

The patch adds functionality to the about:invalidations page, by logging
some calls made on an InvalidatorService implementer (this patch focuses
on Ticl client). Each InvalidatorService now owns an InvalidatorLogger
object that is used to log onInvalidation and onStatusChange.

The invalidationLogger then pushes the invalidation data to the
InvalidationMessageHandler and from there it is sent to the HTML via JS
where it is processed and shown on the textarea as a log file.

Following patches will include a better visualization for the different
types of invalidations, as well as information on the registrars
listening for them.


BUG=263863

Review URL: https://codereview.chromium.org/147353011