If a patch is being applied then we need revert step - take 2.
The following CL was reverted because it broke the gpu bots:
https://codereview.chromium.org/143303003/
The revert: https://codereview.chromium.org/132793013/

That CL broke the GPU bots because they had an empty patch_url property. This CL
should not break the GPU bots, because it checks for a value for patch_url.

BUG=332437

Review URL: https://codereview.chromium.org/153823009