This is the clean-up following r249138.
Introduce a GetUpdatesDelegate class to contain request behavior that is
specific to a GU type.  There are implementations for Normal, Retry,
Poll and Configure GUs.  

These implementations may have different members.  For example, only the
'Normal' GU requires a NudgeTracker.

They also implement different functionality for building the request
message and for applying the received updates.  The latter is to allow
the configure GetUpdates to apply updates directly on the sync thread.

The GU types can still vary in other ways that are not encapsulated in
the GU delegate.  For example, the Poll GU is not retried on failure,
though this is not expressed in its GU delegate in any way.

BUG=278484

Review URL: https://codereview.chromium.org/137753010