Linux Aura: Fix missing system borders on KDE and other window managers.
When a window with system borders does not have a custom shape, its
XShape is set to null, rather than being set to a rectangle the size of
the window. Several window managers (including KDE, XFCE and XMonad)
refuse to put a native window border on a window with a custom shape.

BUG=338916

Review URL: https://codereview.chromium.org/145973004