Add UMA to track app launcher discoverability.
We want to track effectiveness of the app launcher install education
pages in enabling users to find, and start using, the app launcher when
it is auto-enabled upon installing the first v2 packaged app. The
launcher is not auto-shown for the "first" install, to encourage users
to learn where it is.

This CL marks "how" the app launcher is enabled (e.g. command line flag,
webstore, etc.), and when the enable occurs. Then, when the user shows
the app list, the enable method and time taken to show it is recorded.

We also need to record users that "never" show the app list, or who see
the app list show automatically upon installing a _second_ app. One hour
is chosen as a timeout: if no value is recorded after one hour, a value
is recorded in the "last" bucket for the time taken.

BUG=314235
TBR=xiyuan@chromium.org

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=250348

Review URL: https://codereview.chromium.org/143683004