Keep package access members in TOC files
In some cases we have multiple libraries with the same package name
(particularly test+non-test). If one of these depends on the other and
refers to package-private stuff in the second, then it should be rebuilt
if the signature of package-private stuff is changed. To ensure that
happens, the TOC file must include the signatures of package-private
members.

Review URL: https://codereview.chromium.org/161773002