Fix posix IPC channel hanging problem.
If a channel closes right before a send call, listeners might not be notified of
the problem, which can cause hangs. This CL fixes that and adds a test that makes
sure that this does not happen in the future.

This is cl/30133002 + a memory leak fix.

BUG=338709

Review URL: https://codereview.chromium.org/150893002