cc: Simplify picture layer tiling update tile priorities.
This patch simplifies the priority math calculation and moves it
into layers space. This makes the priority to be less correct,
since we're inverse transforming the viewport into layer space
instead of forward transforming tiles into screen space in all cases.

In particular, with perspective and rotated layers, we will be marking
more tiles as visible than before. This should be OK, since the amount
of those layers in practice is limited.

BUG=339142

Review URL: https://codereview.chromium.org/140513006