Add basic coalescing behavior to the permission bubble manager.
The manager will now wait 400ms after it receives both a request and its view
before showing a bubble.  This allows permissions that are requested
quickly in sequence to be shown together, in only one bubble.  Permission
requests that occur when there is already a bubble visible are queued and not
shown until the current bubble disappears.

This simple wait period is probably simpler than the final design, but is a
good stepping stone toward a better algorithm.

BUG=None

Review URL: https://codereview.chromium.org/150103013