Avoid discarding freshly created background tabs
The Chrome OS out-of-memory tab discarder uses the "last selected" time as a signal for which tab to discard. This starts at 0 for freshly created background tabs, like those created with a control-click. The user probably doesn't want to discard those tabs, since they just opened them.

Since the only other user of WebContents::GetLastSelectedTime() is devtools, which uses it for a similar kind of sorting based on activity, rename the method to GetLastActiveTime() and start the value with the tab creation time.

BUG=338735
TEST=content_unittests WebContentsImpl.GetLastActiveTime
TBR=torne@chromium.org for rename touching android_webview

Review URL: https://codereview.chromium.org/138913021