[Galley] Fix an exception in clicking a thumbnail
This patch fixes the exception in clicking a thumbnail. It is a regression caused by r249027 and a possible cause of crbug.com/340624.

The 'length' property of the selection model is NOT the number of the selected item. To get it, we need to get the length of 'selectedIndex' property.

BUG=340624
TEST=manually

Review URL: https://codereview.chromium.org/135973012