Cast: Split kDuplicatePacketReceived into separate events for audio/video.
This change brings this type of event in line with the rest - having separate types
for audio / video. As a new type that is potentially sent over in RTCP is introduced,
a new RTCP encoding for that type is added as well.

This change is backwards-compatible because these events aren't being sent over in RTCP because they are not processed by ReceiverRtcpEventSubscriber.

BUG=337901

Review URL: https://codereview.chromium.org/164253003