Repurpose NotificationBitmapFetcher to BitmapFetcher
There was nothing specific about this class, so I'm generalizing it and moving it upwards in the tree for more use cases (e.g. app banners).

Code that will use it is not ready for review, but is located at https://chromiumcodereview.appspot.com/156343002/

BUG=341556

Review URL: https://codereview.chromium.org/155273002