Disable GpuFeatureTest.MultisamplingAllowed on the official bots
This test is broken because it doesn't take into account that this
feature may be disabled by the blacklist (it is disabled pre-10.8.3 on
basically all vendors). It's only not failing on the main waterfall
because those bots have all features disabled, which the test does
actually handle.

TBR=zmo@chromium.org
BUG=314745
NOTRY=True

Review URL: https://codereview.chromium.org/164983004