Fix automatic toolchain when msvs-ninja in GYP_GENERATORS
Previously, we checked that GYP_GENERATORS was "ninja", but this
broke when using GYP_GENERATORS=ninja,msvs-ninja. As there are no
other supported genearators than those two now, there's no need
to check that we're using ninja at all, so just remove that
branch of the if.

BUG=323300
R=jschuh@chromium.org

Review URL: https://codereview.chromium.org/167373004