[Chromoting] Update jscompile target to use JS files from gypi.
The old version parses the main.html file to get the list of JS files
to verify. Since we're moving to auto-generate this file, we can safely get
the Js files from the same gypi since that's the same source being used
to generate the HTML file.

This also adds the WCS sandbox files to the list of JS files being checked --
they were not being checked as part of the old target because they are not
used in the main.html file.

BUG=
R=jamiewalch@chromium.org

Review URL: https://codereview.chromium.org/163953011