rAc - do less work when model changes
The controller now does a lot more updating/clobbering/fiddling of the view when the model changes (i.e. all the stuff right after view_->ModelChanged()). This means we can afford to do a lot less in AutofillDialogViews::ModelChanged.

Although I'm not sure of the root cause of the crash in the bug, this will also have the side effect of fixing the crash because ModelChanged is part of the stack.

Also, as a further precaution, don't de-ref a possibly NULL |instrument| (even though it does seem like it should definitely be non-NULL).

BUG=343595

Review URL: https://codereview.chromium.org/164543002