GPU: Adjust memory and use the soft-limit.
We now have a soft limit, so we adjust some memory limits up,
and enable the soft-limit to bring the effective usage down.

Initially we bump most limits up a bit (except for really
high-end which is already at 256MB), and then we cut the
soft limit to 50% of that. For low-end, to be extra careful
we only increase the limit from 8MB to 12MB, and set the
soft limit back to 8MB.

BUG=337940,273127
TBR=piman@chromium.org

Review URL: https://codereview.chromium.org/168053002