Support for distilling prior pages in an article.
If distillation is triggered on a page other than the first page, the
distiller only distills the subsequent pages in the article. Add support
for distilling prior pages.
This only adds support to the C++ side, the javascript is still missing
the heuristic to figure out previous pages.

BUG=288015
TEST=Included.

Review URL: https://codereview.chromium.org/167963003