Add temporary define to make some Skia constructors public.
We're currently in the process of enforcing that SkPaint effects can only be
allocated on the heap. That involves making constructors non-public and
providing factory methods instead. Until all the changes on the Skia side have
been rolled into Chrome and the callsites have been changed, we use a define to
guard the visibility of the constructors.

BUG=skia:2187

Review URL: https://codereview.chromium.org/168173003