Remove allow-webui-compositing switch
Long ago, we used to disable compositing on chrome: and about: URLs by default
since that caused a startup performance hit. We then had an --allow-webui-compositing
switch to override this behavior when desired.  The usefulness of this behavior
has long gone and we completely ignore the behavior and the flag on aura builds. This
nukes it from the code

Review URL: https://codereview.chromium.org/167713003