Remove reserved value check in IndependentAndDisposableSamples.
Earlier versions of ISO 14496-12 indicated that the upper 2 bits of
the sample_info was reserved and must be 0. The current version of
ISO 14496-12:2012 now allows these bits to be non-zero so we need to
remove the check so the content will play. The new non-zero values are
not needed by the current code so we don't need any extra logic.

BUG=344737

Review URL: https://codereview.chromium.org/166323005