Clear reserved flags on synthetic continuations.
Previously when WebSocketBasicStream split a frame that arrived in
several chunks, it would copy the reserved bits from the first frame
onto all subsequent frames.

However, the permessage-deflate extension requires that the RSV1 bit
only be set on the first frame of the message.

For compatability with permessage-deflate, clear the reserved bits on
second and subsequent frames when splitting a frame.

BUG=343507
TEST=net_unittests --gtest_filter=WebSocket*

Review URL: https://codereview.chromium.org/171883002