GN: Change gen command syntax, support relative dirs.
This removes the default "gen" command so that "gn" by itself does nothing (it prints a short help message).

The "gen" command now takes as an argument the output directory to use, and I removed the --output flag except for the "gyp" command (which will be needed until we update gyp_chromium).

Updates the base build file to keep in sync with recent changes.

True relative directories are now supported, so that the directories given to the gen command can be relative to the current one. It also supports relative directories for the desc and refs command. This allows you to be in the base directory and say "gn desc :base" and it will resolve the ":base" label relative to the current directory in the source tree.

R=dpranke@chromium.org

Review URL: https://codereview.chromium.org/165823003