Files.app: Use createDirectoryContents_ method during search.
Previously, during typing a search query on the drive, Files.app uses
DirectoryContents#createForDirectory method. It does not handle FakeEntries so
it causes an exception if a user typing a search query in the drive special
folders. This CL replace the method with createDirectoryContents_, which is
a caller of the createForDirectory method and can handle FakeEntries.

BUG=344459
TEST=manually

Review URL: https://codereview.chromium.org/166443008