Telemetry: Look for a page's display name when replaying it.
When recording a page, we identify it by its "display name" (name if it has one,
otherwise URL unless it's a file). When replaying a page, however, we always use
its URL to look up the correct WPR file. So pages with names cannot be found.

R=tonyg@chromium.org
BUG=344380

Review URL: https://codereview.chromium.org/169873002