linux_aura: Keep GTK state from leaking between profiles.
The views::LinuxUI* object is global across profiles. However, it stored
profile specific state in the form of a |use_gtk_| variable. This fixes
Gtk2Borders so that they ask their owning view for a ThemeProvider and
expands the ui::ThemeProvider interface so that it exposes the
UsingNativeTheme() call from the ThemeService layer, instead of asking
the views::LinuxUI provider directly, which will likely return stale state.

BUG=340805, 340799

Review URL: https://codereview.chromium.org/171413002