Leak contexts that are owned by another thread.
This avoids destroying contexts on the main thread that are bound to
the compositor thread as this causes DCHECKs to fire and is a bad thing.

The renderer already does this, as it has been threaded much longer on
more platforms so we noticed it. Recently noticed this in browser tests
in the UI compositor too.

This also updates the Mojo DemoContextFactory to match the behaviour of
GpuProcessTransportFactory wrt the offscreen compositor context creation
and destruction.

R=jamesr, piman
BUG=344016

Review URL: https://codereview.chromium.org/167843002