Unmark newly received resources for deletion.
When a child resource is unused for the frame, but is exported, we mark it for
deletion, pending the return from the top-level parent.
However, we may receive that resource again from the child, and start using it,
before we got the return, so at that point we should unmark the resource for
deletion, because leaving it marked will make us return the resource to the
child when we receive it back from the top-level parent.

BUG=342061

Review URL: https://codereview.chromium.org/170173004