Added early out for pinch to prevent crash when no inner scroll layer.
The layer tree seems to receive gesture events before it's initialized
(or while being destroyed). This means the pinch events get executed
while InnerViewportScrollLayer() == NULL, which was assumed to be non
NULL. The event handler should probably not be sending events to an
uninitialized layer tree but in the mean time I've disabled the
DCHECK and added an early out to prevent crashes.

BUG=340303

Review URL: https://codereview.chromium.org/170743004