Adding new always maximized mode
This is (at this time) not complete. There are some outstanding issues to be resolved marked with TODO(skuhne) which I will address before submitting. This is meant to be a "preview".

What does it do?

It tracks the creation of all browsers / V1 / V2 applications and maximizes them automatically. Non maximizable windows will get centered and a layer will be put behind them. Note that not all windows can be maximized. Dialogs, tray, .. will remain in their window state.

The missing things:
- Background cover layer
- Removing of the restore button from the caption (different CL)
- Suppression of restore / fullscreen request events of any of the tracked windows.

BUG=337563
TEST=unit tests

Review URL: https://codereview.chromium.org/167503002