Correctly handle any number of channels in WebAudio file decoder on Android.
CopyPcmDataToBus incorrectly assumed that each pipe contained a
integer number of frames.  This is only true if the number of channels
is 1 or 2.  For more channels, this is false, so we have to copy the
data carefully, keeping track of which sample in a frame we are in.

BUG=343648

Review URL: https://codereview.chromium.org/164023007