Fixed rounding issue on scrollbar rasterization.
Changed the painting method to translate the canvas back to the
origin before performing the scale to prevent losing pixels due
to rounding. Also made the it take the layer-space rect as a 
parameter, instead of trying to reconstruct it from the 
(possibly-clamped) content-rect, and calculate the layer-space
to content-space scaling factor directly from rects.

BUG=336534

Review URL: https://codereview.chromium.org/150603004