Removing .html reference in Pepper API referenc
The index page that links to Pepper C and Pepper C++ reference
includes .html in links. I've removed this as a precautionary step.

The cache needs clearing so that we don't link to .rst in it.

To speed things up and make sure this is clean-- I'm removing .html to be safe.

NOTRY=true

Review URL: https://codereview.chromium.org/174363002