Fix APK dependency on remoting resources
This ensures that Ant regenerates the R.java resource file if the
remoting/resources/remoting_strings.grd file is changed. This had the
potential to cause incremental build failures if a new Android string
were added to this grd file (and a .java file tried to access the new
resource).

TEST=Build remoting_apk, touch the remoting grd file and rebuild the target - verify that the R.java/R.txt step runs.
R=jamiewalch@chromium.org

Review URL: https://codereview.chromium.org/171423018