Fix posix IPC channel hanging problem.
If a channel closes right before a send call, listeners might not be notified of
the problem, which can cause hangs. This CL fixes that and adds a test that makes
sure that this does not happen in the future.

This is similar to cl/150893002, but takes a slightly different approach to how to
make sure everything happens in the right order. In particular, it avoids closing
the socket (and calling OnChannelError()) from Send().

BUG=338709

Review URL: https://codereview.chromium.org/172773002