Introduce Workspace related event, and moved the code to adjust bounds from WLM to DefaultState.
Cleaned up the code used when the window is added to the workspace. We used to try to maximize twice (depending on the code path).

Added test case for the case where the window with an empty bounds and maximzied state is added to the workspace.

* Next step: Make WMEvent into classes.

BUG=318325
TEST=WorkspaceLayoutManagerTest.MaximizeWithEmptySize

Review URL: https://codereview.chromium.org/138013008