Perf: Associate page set with existing WPR
The tough_animation_cases page set contains the live site
http://ie.microsoft.com/testdrive/performance/robohornetpro/
but has not specified a WPR of the site for perf bots to use.

This patch associates this page set with an existing WPR of the site used
by other benchmarks for the same site.


BUG=309981

Review URL: https://codereview.chromium.org/173353002