Copy the * gypi OWNERS rule from chrome/ to content/
Many directories under content/ have specific OWNERS for that topic area,
but all of the gypi rules live in content/ because of restrictions with gypi
and the component build. It's not a good use of time to find a separate OWNER
for a file addition/removal/rename when a patch already has a topic-appropriate
owner. This copies the *.gypi * OWNERS rule from chrome/OWNERS with the same comment.

Review URL: https://codereview.chromium.org/173833005