ui: Allow individual test suites to opt-out of the TestCompositor.
This adds a temporary bool to GLSurface::InitializeOneOffForTests()
allowing a test suite to opt-in to NullDraw bindings. When a test
suite as opted in, the InitializeContextFactoryForTests() method
can see, and will not use the TestContextFactory, instead using the
InProcessContextFactory.

Once all test suites are optign into NullDraw bindings. The bool
and query method for HasInitializedNullDrawGLBindings() will be
removed.

R=kbr@chromium.org, sievers@chromium.org
BUG=270918

Review URL: https://codereview.chromium.org/175093004