Refactor user pods to use authType property for distinct authentication modes.
The authType of the user pod determines the authentication method to sign in
or unlock from the account picker screen. This refactoring aims to unify the
current two modes the user pod can use, and make the user pod more
extensible for other modes.

For example, the user pod may currently authenticate using a password or online
GAIA signin. This CL introduces a full-pod user click authType, and a PIN
code authType will be implemented in the future (see bug). This CL also adds
plumbing for the authType to be set by the ScreenLocker for the
chrome.screenlockPrivate API.

BUG=344179
TEST=visual inspection

Review URL: https://codereview.chromium.org/168813002