Remove default browser uma stat.
This metric has been broken since the day it was introduced, over a year
ago, suggesting that nobody cares about it anyway, hence the decision to
remove it rather than fix it.

The actual enum for this is in internal histograms.xml, so that will be
uploaded as a separate review.

BUG=312343
R=isherman@chromium.org

Review URL: https://codereview.chromium.org/175373004