[Android] Allow inspection of JavaScript <-> Java bridge objects
This change allows enumeration of methods of JavaScript objects added
via "addJavascriptInterface" by means of JS "Object.keys()" method and
"for .. in" loop. This is mostly useful for debugging hybrid Android apps,
as DevTools can now list and suggest the names of the methods that
the object provides.

This feature is made configurable, and is disabled by default due to
compatibility risks. It will be enabled for Android applications
targeting newer versions of API.

BUG=333258

Review URL: https://codereview.chromium.org/135653015