Replaced HEAD with origin/master for "revision" parameter value
polymer-dev builder used HEAD as a revision parameter value for
gclient sync. The following command was executed:

  git rebase --onto HEAD origin/master master

which is a NOOP.

AFAIU, from the gclient-sync perspective, "HEAD" revision doesn't make
sense because it must be a revision that a remote is aware of.
Going to upload a patch to gclient to throw an error if revision=HEAD

R=agable@chromium.org
BUG=333454

Review URL: https://codereview.chromium.org/175663002