Fixes minor glitches of the sound level indicator circle.
The existing code sets the center of the circle at the center point
of the mic image. However, the mic circle is positioned slightly
above in the image (to draw the shadow at the bottom).

This CL adds the offset for the vertical position of the indicator
circle and the offset for the minimum radius.

Also, sets the anti-alias flag for painting the circle to draw
the edge smoothly.

BUG=None
R=xiyuan@chromium.org
TEST=my eyes

Review URL: https://codereview.chromium.org/168243004