cc: Unified the implementation of running tasks on origin and worker threads.
RasterWorkerPoolTaskImpl::RunOnWorkerThread and RasterWorkerPoolTaskImpl::RunOnOriginThread are very similar now. The only difference is we do not need to clone the picture-pile when running on origin thread. This will need to be revisited when we actually turn on hybrid or multi-threaded rasterization. I also moved push-pop GL marker to DirectRasterWorkerPool, which removed the dependency of RasterWorkerPoolTaskImpl on ContextProvider.

Review URL: https://codereview.chromium.org/174613004