aura: Make Window::HitTest() a private method.
This is the first step towards removing Window::HitTest(), and related methods
from WindowDelegate (and corresponding methods in views::WidgetDelegate). This
is currently used only in a handful of places. This CL replaces those calls
with equivalent code, and makes Window::HitTest() a private method, so that
new code do not use it.

BUG=318879
R=sky@chromium.org

Review URL: https://codereview.chromium.org/178493003