Fix the extra movement issue when an app list item is dropped to another app list item to create a new folder.
When the new item view is created for the new folder, its bounds is set to (0,0) by default, therefore, creating the extra movement when AninamateToIdealBounds is called to animate it from its current bounds to the ideal bounds. Fixed now by setting its bounds to the same as the old item view it is going to replace.

BUG=343354
R=xiyuan@chromium.org

Review URL: https://codereview.chromium.org/165723013