Fixing the GCMStoreImplTest.GetNextPersistentId on Windows
Problem:
It seems that consecutive calls to base::Time::Now().ToInternalValue() might
yield the same outcome multiple times in a row, which is not suitable for
producing unique IDs for testing.

Fix:
Persist a starting value based on current time and then increment it with
every call.

BUG=345118
R=zea@chromium.org

Review URL: https://codereview.chromium.org/177243007