[OriginChip] Set origin_chip_enabled to true in ShowURL().
This wasn't necessary in my original show/hide implementation which set this value to true on any non-click focus.  But when we got rid of that, this left the ShowURL case (called by clicking the chip or Ctrl-l/Alt-d) uncovered.  You don't notice it at first because url_replacement_enabled is set to false, which supresses the chip, but that get's set back to true when the user starts editing.  This change ensures that the chip remains hidden during editing and is then reshown of focus loss.

BUG=338563

Review URL: https://codereview.chromium.org/177413007