Isolate all scrolling hacks in the LegacyRenderWidgetHostHWND class.
This includes setting the WS_VSCROLL and WS_HSCROLL styles on the HWND, etc. Sadly we need both the matching class
name and the styles. As some drivers like the USB driver check the styles as well.

In any case this is on the same lines as code in WebContentsViewWin where we set these styles and returned 0 from WM_NCCALCSIZE
to prevent Windows from drawing them.

Temporarily disabled the code in HWNDMessageHandler which adds scroll styles for top level windows. Will take the whole code
out when we hear back from Canary users about whether this is working.

BUG=335941
R=cpu@chromium.org, sky@chromium.org, cpu, sky

Review URL: https://codereview.chromium.org/178183003