Use validating IPC enum macros in accessibility messages.
This is part of a long-running background task to remove the remaining uses of the unchecked IPC_ENUM_TRAITS() macro.

The most automated way to do this is to tweak the json_schema_compiler to generate these for us. We'd need to do something as currently the idl doesn't support providing values for the corresponding enum

BUG=246708

Review URL: https://codereview.chromium.org/143873021