Mojo: Attach dispatchers to message earlier.
In particular, ProxyMessagePipeEndpoint::EnqueueMessage() gets a message with
dispatchers already attached. It will be up to the Channel to force them into a
serialized state.

Note that this is a bit less optimal (especially right now), since we means we
have to create duplicate dispatchers even for things that will be serialized
immediately. However, this structure makes things more flexible, since we can
defer serialization if we want (e.g., we don't want Channel to send everything
to RawChannel in a serialized state immediately).

R=yzshen@chromium.org

Review URL: https://codereview.chromium.org/177073015