Web MIDI: link with alsa even if cras is enabled
This CL fixes a feature breakage on Chrome OS which happens by
crrev.com/244251 accidentally.

Chrome for Chrome OS should be linked with alsa even if cras is enabled.
This is because Web MIDI implementation for Chrome OS depends on alsa.

Also this changes add trace events for each platform implementation.
This is also useful to know which implementation is enabled in the build.

BUG=335865

Review URL: https://codereview.chromium.org/171603014