Refactor TokenValidatorImpl into a base class + implementation.
Most of the common logic and response handling code are moved into a base class, and the implementation contains just the logic to prepare the validation request (which is where the actual keypair-based authentication happens). This makes it easier to implement different, non-keypair-based host authentication mechanisms.

BUG=

Review URL: https://codereview.chromium.org/165293004